data-generator/requirements.txt
kafka-python==2.0.2
faker==11.3.0
python-dateutil==2.8.2
